master_backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
protobuf==5.27.0
aiofiles==23.2.1
python-dotenv==1.0.1
zeroconf==0.132.2
paramiko==3.4.0
scipy==1.13.1
